requirements.txt
absl-py==2.2.1
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
dill==0.3.8
-e git+https://github.com/unfiltered-syrup/English_To_ASL_Renderer.git@a937fa51d4a8033409817f8efd431cd509de9f96#egg=English_To_ASL_Renderer
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2024.12.0
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.3
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==3.1.1
propcache==0.3.1
protobuf==4.25.6
proxy-tools==0.1.0
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.2
pywebview==5.4
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
